# api/openapi.yaml
openapi: 3.0.3
info:
  title: ZJC 航空公司机票预订系统 API
  description: ZJC 航空公司机票预订系统 REST API 接口文档
  version: 1.0.0
  contact:
    name: API Support
    email: support@zjc-airline.com
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
  - url: http://api.zjc-airline.com
    description: 生产服务器
...
paths:
...
  /api/flights:
    get:
      tags:
        - 航班管理
      summary: 获取所有航班
      description: 获取系统中所有航班列表
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseFlightList'
    post:
      tags:
        - 航班管理
      summary: 创建航班
      description: 管理员创建新航班
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flight'
            example:
              flightNumber: "ZJC016"
              aircraftType: "波音737-800"
              origin: "北京"
              destination: "广州"
              departureTime: "2025-12-06T08:00:00"
              arrivalTime: "2025-12-06T11:30:00"
              price: 950.00
              remainingSeats: 30
              remarks: "经济舱，含餐食"
      responses:
        '201':
          description: 创建成功
...
  /api/flights/{id}:
    get:
      tags:
        - 航班管理
      summary: 根据ID获取航班
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
    put:
      tags:
        - 航班管理
      summary: 更新航班信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flight'
      responses:
        '200':
          description: 更新成功
    delete:
      tags:
        - 航班管理
      summary: 删除航班
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功
  /api/flights/number/{flightNumber}:
    get:
      tags:
        - 航班管理
      summary: 根据航班号获取航班
      parameters:
        - name: flightNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
  /api/flights/search:
    post:
      tags:
        - 航班管理
      summary: 搜索航班
      description: 根据出发地、目的地搜索航班
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlightSearchRequest'
            example:
              origin: "北京"
              destination: "上海"
      responses:
        '200':
          description: 成功
  /api/bookings:
    get:
      tags:
        - 预订管理
      summary: 获取所有预订
      description: 获取系统中所有预订记录
      responses:
        '200':
          description: 成功
    post:
      tags:
        - 预订管理
      summary: 创建预订
      description: 创建新的机票预订
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
            example:
              flightId: 1
              userId: 1
              ticketCount: 2
      responses:
        '201':
          description: 创建成功
  /api/bookings/{id}:
    get:
      tags:
        - 预订管理
      summary: 根据ID获取预订
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
    delete:
      tags:
        - 预订管理
      summary: 删除预订
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功
...
  /api/bookings/flight/{flightId}:
    get:
      tags:
        - 预订管理
      summary: 获取航班的预订列表
      parameters:
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
  /api/bookings/{id}/status:
    put:
      tags:
        - 预订管理
      summary: 更新预订状态
      description: 更新预订状态（PAID, CANCELED等）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [CREATED, PAID, CANCELED]
                  example: "PAID"
      responses:
        '200':
          description: 更新成功
components:
  schemas:
...
    BookingRequest:
      type: object
      required:
        - flightId
        - userId
        - ticketCount
      properties:
        flightId:
          type: integer
          format: int64
          description: 航班ID
        userId:
          type: integer
          format: int64
          description: 用户ID
        ticketCount:
          type: integer
          description: 票数
    FlightSearchRequest:
      type: object
      properties:
        origin:
          type: string
          description: 出发地
        destination:
          type: string
          description: 目的地
        departureDate:
          type: string
          format: date-time
          description: 出发日期
...
    Flight:
      type: object
      properties:
        id:
          type: integer
          format: int64
        flightNumber:
          type: string
        aircraftType:
          type: string
        origin:
          type: string
        destination:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        price:
          type: number
          format: decimal
        remainingSeats:
          type: integer
        remarks:
          type: string
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
...
    ApiResponseFlightList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Flight'
...